Modify model/signUp.py
